requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
huggingface-hub==0.24.6
HyperPyYAML==1.2.2
idna==3.7
intel-openmp==2021.4.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
protobuf==4.25.4
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.4.0
speechbrain==1.0.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
sympy==1.13.2
tabulate==0.9.0
tbb==2021.13.1
tensorboardX==2.6.2.2
textstat==0.7.4
threadpoolctl==3.5.0
tokenizers==0.19.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchmetrics==1.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4